:recycle: Update test directory structure to identify protocol specs
