--- elements/lisk-cryptography/test/eth2_bls_specs
@@ -1,1 +1,0 @@
-../../../protocol-specs/eth2_spec_tests/tests/general/phase0/bls
