+++ elements/lisk-cryptography/test/helpers/index.ts
@@ -12,9 +12,37 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
+
+import * as glob from 'glob';
+import { join } from 'path';
+import * as fs from 'fs';
+import * as yml from 'js-yaml';
+
+export const EMPTY_BUFFER = Buffer.alloc(0);
+
 export const makeInvalid = (buffer: Buffer): Buffer => {
 	const replace = buffer[0] % 2 === 0 ? 1 : 2;
 	// eslint-disable-next-line no-param-reassign
 	buffer[0] = replace;
 	return buffer;
 };
+
+export const getAllFiles = (
+	dirs: string[],
+	ignore?: RegExp,
+): { path: string; toString: () => string }[] => {
+	return dirs
+		.map((dir: string) => {
+			return glob
+				.sync(join(__dirname, '../protocol_specs/', dir, '**/*.{yaml,yml}'))
+				.filter(path => (ignore ? !ignore.test(path) : true))
+				.map(path => ({ path, toString: () => `${dir}${path.split(dir)[1]}` }));
+		})
+		.flat();
+};
+
+export const loadSpecFile = <T = Record<string, unknown>>(filePath: string) =>
+	(yml.load(fs.readFileSync(filePath, 'utf8')) as unknown) as T;
+
+export const hexToBuffer = (str: string | null): Buffer =>
+	str ? Buffer.from(str.substr(2), 'hex') : EMPTY_BUFFER;
