+++ elements/lisk-cryptography/test/protocol_specs/bls_specs/aggregate/one_signature_no_e2_curve.yml
