+++ elements/lisk-cryptography/test/protocol_specs/bls_specs/aggregate/third_signature_not_a_point_on_E2.yml
