+++ elements/lisk-cryptography/test/protocol_specs/bls_specs/create_agg_sig/case2.yml
