+++ elements/lisk-cryptography/test/protocol_specs/bls_specs/pop_prove/case3.yml
