+++ elements/lisk-cryptography/test/protocol_specs/bls_specs/pop_verify/pk_in_E1_but_not_in_G1.yml
