+++ elements/lisk-cryptography/test/protocol_specs/bls_specs/pop_verify/pk_is_identity_point.yml
