+++ elements/lisk-cryptography/test/protocol_specs/bls_specs/pop_verify/tampered_proof.yml
