+++ elements/lisk-cryptography/test/protocol_specs/bls_specs/sign_bls/case1.yml
