+++ elements/lisk-cryptography/test/protocol_specs/bls_specs/sk_to_pk/case2.yml
