+++ elements/lisk-cryptography/test/protocol_specs/bls_specs/sk_to_pk/case3.yml
