+++ elements/lisk-cryptography/test/protocol_specs/bls_specs/verify_bls/case1.yml
