+++ elements/lisk-cryptography/test/protocol_specs/bls_specs/verify_weighted_agg_sig/case2.yml
