+++ elements/lisk-cryptography/test/protocol_specs/eth2_bls_specs
@@ -1,0 +1,1 @@
+../../../../protocol-specs/eth2_spec_tests/tests/general/phase0/bls
